Fix unused imports
Remove unused imports flagged by the Rust compiler across multiple crates:

- lib/typed_indexmap/src/set.rs
- openvaf/hir_lower/src/callbacks.rs
- openvaf/mir/src/*.rs (cursor, dfg, flowgraph, layout, lib, etc.)
- openvaf/mir_autodiff/src/builder.rs
- openvaf/mir_build/src/lib.rs
- openvaf/mir_llvm/src/builder.rs
- openvaf/mir_opt/src/global_value_numbering.rs
- openvaf/osdi/src/*.rs (bitfield, eval, inst_data, lib, load, etc.)
- openvaf/sim_back/src/context.rs
- verilogae/verilogae/src/compiler_db.rs

+++ openvaf/osdi/src/compilation_unit.rs
@@ -56,7 +56,7 @@ pub fn new_codegen<'a, 'll>(
     for fun in function_iter(llmod.llmod()) {
         unsafe {
             // LLVMPurgeAttrs(fun);
-            if LLVMIsDeclaration(fun) != 0 as i32 {
+            if LLVMIsDeclaration(fun) != 0_i32 {
                 continue;
             }
